Home page changes
Section1 :
Please change particles.js to not look like it’s attacking the curser,
something closer to the original state
Particles.js shouldn’t be parallax, it should stay with the top section,
make sure it has the same boundaries as the entire top section (that it
extends to the bottom of the picture).

Section2:
Add particles_section2.png to top of section (like original mock).
Make the boxes and text in boxes wider so the boxes won’t be so long.
Add padding below “Approach”, above boxes

Section3:
Center the “Discover what’s possible...” and Learn more button in the
section (add more padding above text)

Section4:
Move testimonial arrows to the respective left and right sides.
Increase the amount of time each testimonial is on screen (slow it down)
center everything into the box